rename preset components
